pyproject.toml
[tool.poetry]
name = "attendance-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "attendance_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
channels = "3.0.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
